requirements.txt
PixivPy==3.7.0
fastapi==0.79.1
uvicorn==0.18.2
requests==2.28.1
